removed a bunch of commands to make life easier
Removed PI_CTRL_SYN_ACK, PI_CTRL_BUSY, PI_CTRL_SYN, PI_CTRL_DISCONNECT, and PI_CTRL_SET_NAME. The reason for removing SYN, SYN_ACK, BUSY, and DISCONNECT are because I'll let the sockets handle all of that. The effect is the same (and less work to do) if we just set the backlog # of connections to 1. We will sacrifice SET_NAME for this, but it'll make everyone's lives easier